rcd: move webgui apart; option to disable browser
Fix #3601, #3785
